requirements.txt
fastapi>=0.114.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
openai>=1.51.0
anyio>=4.4.0
pytest>=8.2.0
httpx>=0.27.0
